NDEV-2332: Decouple tracer code from evm-loader to neon-lib crate (#232)
* NDEV-2332: Move tracer impl from evm_loader to neon_lib

* NDEV-2332: Move types from program::evm::tracing::mod to lib::tracing::mod

* NDEV-2332: Remove Send + Sync from EventListener

* NDEV-2332: Remove unused impl

* NDEV-2332: Convert tracing module from directory to file
