+++ evm_loader/lib/src/tracing/tracers/struct_logger.rs
@@ -1,14 +1,14 @@
 use std::collections::BTreeMap;
 
-use crate::evm::ExitStatus;
 use ethnum::U256;
+use evm_loader::evm::ExitStatus;
 use serde::Serialize;
 use serde_json::{json, Value};
 
-use crate::evm::opcode_table::OPNAMES;
-use crate::evm::tracing::TraceConfig;
-use crate::evm::tracing::{Event, EventListener};
-use crate::types::hexbytes::HexBytes;
+use crate::tracing::TraceConfig;
+use evm_loader::evm::opcode_table::OPNAMES;
+use evm_loader::evm::tracing::{Event, EventListener};
+use evm_loader::types::hexbytes::HexBytes;
 
 /// `StructLoggerResult` groups all structured logs emitted by the EVM
 /// while replaying a transaction in debug mode as well as transaction
